feature(accessTokenManagement): make the project trimmable
AccessTokenManagement is just using a couple serialization api that aren't trim and aot compatible which can be handled by source generation.

Technically the Dictionary<string, object> require string to be added to the list but since ClientCredentialsToken already has string children I did not include it to save code. Only the long from the date time conversion is required for tests to pass.