Simplify ProtocolResponse.FromHttpResponseAsync
We now always invoke InitializeAsync at the end of the method.
This simplifies the code path, especially for jwt introspection, where we
want to take control of the parsing of the json within the jwt in the derived
TokenIntrospectionResponse instead of in ProtocolResponse.

We can now pass skipJsonParsing: true, and then deal with initialization ourselves.